Separate out header files and add more explanation in the C++ header
