+++ src/ort_genai_c.h
@@ -237,165 +237,4 @@ OGA_EXPORT OgaResult* OGA_API_CALL OgaGetCurrentGpuDeviceId(int* device_id);
 
 #ifdef __cplusplus
 }
-
-// This is added as a member variable in the wrapped types to prevent accidental construction/copying
-// Since the wrapped types are never instantiated by value, this member doesn't really exist. The types are still opaque.
-struct OgaAbstract {
-  OgaAbstract() = delete;
-  OgaAbstract(const OgaAbstract&) = delete;
-  void operator=(const OgaAbstract&) = delete;
-};
-
-struct OgaResult : OgaAbstract {
-  static void operator delete(void* p) { OgaDestroyResult(reinterpret_cast<OgaResult*>(p)); }
-  const char* GetError() const { return OgaResultGetError(this); }
-};
-
-inline void OgaCheckResult(OgaResult* result) {
-  if (result) {
-    std::unique_ptr<OgaResult> p_result{result};  // Take ownership so it's destroyed properly
-    throw std::runtime_error(p_result->GetError());
-  }
-}
-
-struct OgaModel : OgaAbstract {
-  static std::unique_ptr<OgaModel> Create(const char* config_path) {
-    OgaModel* p;
-    OgaCheckResult(OgaCreateModel(config_path, &p));
-    return std::unique_ptr<OgaModel>(p);
-  }
-
-  std::unique_ptr<OgaSequences> Generate(const OgaGeneratorParams& params) {
-    OgaSequences* p;
-    OgaCheckResult(OgaGenerate(this, &params, &p));
-    return std::unique_ptr<OgaSequences>(p);
-  }
-
-  static void operator delete(void* p) { OgaDestroyModel(reinterpret_cast<OgaModel*>(p)); }
-};
-
-struct OgaString {
-  OgaString(const char* p) : p_{p} {}
-  ~OgaString() { OgaDestroyString(p_); }
-
-  operator const char*() const { return p_; }
-
-  const char* p_;
-};
-
-struct OgaSequences : OgaAbstract {
-  static std::unique_ptr<OgaSequences> Create() {
-    OgaSequences* p;
-    OgaCheckResult(OgaCreateSequences(&p));
-    return std::unique_ptr<OgaSequences>(p);
-  }
-
-  size_t Count() const {
-    return OgaSequencesCount(this);
-  }
-
-  std::span<const int32_t> Get(size_t index) const {
-    return {OgaSequencesGetSequenceData(this, index), OgaSequencesGetSequenceCount(this, index)};
-  }
-
-  static void operator delete(void* p) { OgaDestroySequences(reinterpret_cast<OgaSequences*>(p)); }
-};
-
-struct OgaTokenizer : OgaAbstract {
-  static std::unique_ptr<OgaTokenizer> Create(const OgaModel& model) {
-    OgaTokenizer* p;
-    OgaCheckResult(OgaCreateTokenizer(&model, &p));
-    return std::unique_ptr<OgaTokenizer>(p);
-  }
-
-  void Encode(const char* str, OgaSequences& sequences) const {
-    OgaCheckResult(OgaTokenizerEncode(this, str, &sequences));
-  }
-
-  OgaString Decode(std::span<const int32_t> tokens) const {
-    const char* p;
-    OgaCheckResult(OgaTokenizerDecode(this, tokens.data(), tokens.size(), &p));
-    return p;
-  }
-
-  static void operator delete(void* p) { OgaDestroyTokenizer(reinterpret_cast<OgaTokenizer*>(p)); }
-};
-
-struct OgaTokenizerStream : OgaAbstract {
-  static std::unique_ptr<OgaTokenizerStream> Create(const OgaTokenizer& tokenizer) {
-    OgaTokenizerStream* p;
-    OgaCheckResult(OgaCreateTokenizerStream(&tokenizer, &p));
-    return std::unique_ptr<OgaTokenizerStream>(p);
-  }
-
-  /*
-   * Decode a single token in the stream. If this results in a word being generated, it will be returned in 'out'.
-   * The caller is responsible for concatenating each chunk together to generate the complete result.
-   * 'out' is valid until the next call to OgaTokenizerStreamDecode or when the OgaTokenizerStream is destroyed
-   */
-  const char* Decode(int32_t token) {
-    const char* out;
-    OgaCheckResult(OgaTokenizerStreamDecode(this, token, &out));
-    return out;
-  }
-
-  static void operator delete(void* p) { OgaDestroyTokenizerStream(reinterpret_cast<OgaTokenizerStream*>(p)); }
-};
-
-struct OgaGeneratorParams : OgaAbstract {
-  static std::unique_ptr<OgaGeneratorParams> Create(const OgaModel& model) {
-    OgaGeneratorParams* p;
-    OgaCheckResult(OgaCreateGeneratorParams(&model, &p));
-    return std::unique_ptr<OgaGeneratorParams>(p);
-  }
-
-  void SetSearchOption(const char* name, int value) {
-    OgaCheckResult(OgaGeneratorParamsSetSearchNumber(this, name, value));
-  }
-
-  void SetSearchOption(const char* name, double value) {
-    OgaCheckResult(OgaGeneratorParamsSetSearchNumber(this, name, value));
-  }
-
-  void SetSearchOption(const char* name, bool value) {
-    OgaCheckResult(OgaGeneratorParamsSetSearchBool(this, name, value));
-  }
-
-  void SetInputIDs(const int32_t* input_ids, size_t input_ids_count, size_t sequence_length, size_t batch_size) {
-    OgaCheckResult(OgaGeneratorParamsSetInputIDs(this, input_ids, input_ids_count, sequence_length, batch_size));
-  }
-
-  void SetInputSequences(const OgaSequences& sequences) {
-    OgaCheckResult(OgaGeneratorParamsSetInputSequences(this, &sequences));
-  }
-
-  static void operator delete(void* p) { OgaDestroyGeneratorParams(reinterpret_cast<OgaGeneratorParams*>(p)); }
-};
-
-struct OgaGenerator : OgaAbstract {
-  static std::unique_ptr<OgaGenerator> Create(const OgaModel& model, const OgaGeneratorParams& params) {
-    OgaGenerator* p;
-    OgaCheckResult(OgaCreateGenerator(&model, &params, &p));
-    return std::unique_ptr<OgaGenerator>(p);
-  }
-
-  bool IsDone() const {
-    return OgaGenerator_IsDone(this);
-  }
-
-  void ComputeLogits() {
-    OgaCheckResult(OgaGenerator_ComputeLogits(this));
-  }
-
-  void GenerateNextToken() {
-    OgaCheckResult(OgaGenerator_GenerateNextToken(this));
-  }
-
-  std::span<const int32_t> GetSequence(size_t index) const {
-    return {OgaGenerator_GetSequence(this, index), OgaGenerator_GetSequenceLength(this, index)};
-  }
-
-  static void operator delete(void* p) { OgaDestroyGenerator(reinterpret_cast<OgaGenerator*>(p)); }
-};
-
 #endif
